Game: remove keyboardinput and window resizing in new files
